# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/carpark-info": {
      "get": {
        "operationId": "ping",
        "summary": "",
        "description": "Health Check",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health Check"
        ]
      }
    },
    "/api/carpark-info/auth/sign-in": {
      "post": {
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthSignInRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/carpark-info/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRegister"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/carpark-info/auth/refresh-token": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRefreshToken"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/carpark-info/user/me": {
      "get": {
        "operationId": "UserController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/carpark-info/carpark": {
      "get": {
        "operationId": "CarparkController_getList",
        "parameters": [
          {
            "name": "freeParking",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "YES",
                "NO"
              ],
              "type": "string"
            }
          },
          {
            "name": "nightParking",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "YES",
                "NO"
              ],
              "type": "string"
            }
          },
          {
            "name": "gantryHeight",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Carpark"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/carpark-info/favorite-list": {
      "post": {
        "operationId": "FavoriteListController_addNewFavoriteCarpark",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddFavoriteCarpark"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Favorite List"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "FavoriteListController_getMyFavoriteCarpark",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Favorite List"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/carpark-info/favorite-list/{id}": {
      "delete": {
        "operationId": "FavoriteListController_deleteAFavoriteCarpark",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Favorite List"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Carpark Info API",
    "description": "API specification for Carpark Info Service | [swagger.json](swagger.json)",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "AuthSignInRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "AuthRegister": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password",
          "fullName"
        ]
      },
      "AuthRefreshToken": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "AddFavoriteCarpark": {
        "type": "object",
        "properties": {
          "carParkNo": {
            "type": "string"
          }
        },
        "required": [
          "carParkNo"
        ]
      }
    }
  }
}